requirements.txt
attrs==22.2.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
psycopg2==2.9.10
pytest==7.2.2
python-dotenv==1.0.0
tabulate==0.9.0

pandas==2.2.3